+++ src/gamslib/objectcsv/defaultvalues.py
@@ -1,6 +1,6 @@
 """Default values for the datastream meatadata."""
 
-import xml.etree.ElementTree as ET
+
 NAMESPACES = {
     "tei": "http://www.tei-c.org/ns/1.0",
     "lido": "http://www.lido-schema.org",
@@ -20,27 +20,5 @@ FILENAME_MAP = {
         "title": "Dublin Core Metadata",
         "description": "Dublin Core Metadata in XML format for this content file.",
     },
-    # Elisabeth says, this makes no sense; better extract from TEI/LIDO
-    #"TEI.xml": {
-    #    "title": "Main TEI file",
-    #    "description": "The central TEI File for this object",
-    #},
-    #"LIDO.xml": {
-    #    "title": "Main LIDO file",
-    #    "description": "The central LIDO file of this object",
-    #},
     "RDF.xml": {"title": "RDF Statements", "description": ""},
 }
-
-
-def extract_title_from_tei(tei_file):
-    tei = ET.parse(tei_file)
-    title_node = tei.find('tei:teiHeader/tei:fileDesc/tei:titleStmt/tei:title', namespaces=NAMESPACES)
-    return title_node.text if title_node is not None else ""
-
-
-def extract_title_from_lido(lido_file):
-    "Extract the title from a LIDO file."
-    lido = ET.parse(lido_file)
-    title_node = lido.find('lido:descriptiveMetadata/lido:objectIdentificationWrap/lido:titleWrap/lido:titleSet/lido:appellationValue', namespaces=NAMESPACES)
-    return title_node.text if title_node is not None else ""