+++ src/gamslib/objectcsv/utils.py
@@ -1,10 +1,11 @@
 """Utility functions for the objectcsv module."""
 
 import logging
+import warnings
+import xml.etree.ElementTree as ET
 from pathlib import Path
 from typing import Generator
-import warnings
-
+from .defaultvalues import NAMESPACES
 logger = logging.getLogger()
 
 
@@ -19,3 +20,15 @@ def find_object_folders(root_directory: Path) -> Generator[Path, None, None]:
                     f"Skipping '{directory}' as folder does not contain a DC.xml file.",
                     UserWarning
                 )
+
+def extract_title_from_tei(tei_file):
+    tei = ET.parse(tei_file)
+    title_node = tei.find('tei:teiHeader/tei:fileDesc/tei:titleStmt/tei:title', namespaces=NAMESPACES)
+    return title_node.text if title_node is not None else ""
+
+
+def extract_title_from_lido(lido_file):
+    "Extract the title from a LIDO file."
+    lido = ET.parse(lido_file)
+    title_node = lido.find('lido:descriptiveMetadata/lido:objectIdentificationWrap/lido:titleWrap/lido:titleSet/lido:appellationValue', namespaces=NAMESPACES)
+    return title_node.text if title_node is not None else ""