+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2025-01-30
+
+### Added
+
+* A new subpackage formatdetect for file format identification 
+
+### Changed
+
+* project.toml has 2 new fields in the general table: format_detector and format_detector_url
+* Add tags column to datastreams.csv
+* Improve auto-detection for some csv columns
+* The Configuration object in submodule projectconfiguration now must be obtained by calling
+  the (cached) get_configuration() function
 
 ## [0.3.1] - 2025-01-09